requirements.txt
flask==3.1.0
flask-restx==1.3.0
sqlalchemy==2.0.36
uuid==1.30
pymongo==4.10.1
dependency-injector==4.44.0
pymongo==4.10.1
python-dotenv==1.0.1